feat: extract DataLoadingOrchestrator component from DataManager
Phase 3 of DataManager reduction project - extract data loading orchestration
logic to dedicated service component for functionality-preserving refactoring.

## Key Changes
- **NEW**: DataLoadingOrchestrator class (474 lines)
- **MODIFIED**: DataManager reduced from 1,675 to 1,164 lines (-511 lines)
- **REDUCTION**: 30% size reduction achieved in Phase 3

## Architecture Improvements
- Clean separation of data loading orchestration concerns
- Loose coupling via dependency injection pattern
- Preserved exact functionality with enhanced maintainability
- Enhanced mode integration (tail/backfill/full) through orchestrator

## Quality Assurance
- All existing DataManager tests pass
- End-to-end functionality confirmed working
- No behavioral changes or regressions
- Clean architectural boundaries maintained

## Implementation Details
- Extracted _load_with_fallback() method (442 lines) to orchestrator
- Used DataLoadingOrchestrator(self) pattern for loose coupling
- Maintained all existing method signatures and return types
- Preserved complex orchestration logic exactly as-is

Progress: DataManager 2,055 → 1,164 lines (target <500 lines)
